#6: Relocate `SettingsMenu` screen to /Settings folder
